Fix combinations check
In thirty bees 1.4.0, method Product::getAttributesResume returns empty
array for products without combinations. Unfortunately that is braking
change, and current check in the module

is_array($attributesResume) > 0

returns true, instead of expected false.

The check is quite strange, I suspect the original code used to be

count($attributesResume) > 0

but was changed because it raised warnings in PHP7.

I propose to change the check simply to to

$attributesResume

Related to thirtybees/thirtybees#1467